Добавил два эндпоинта для удаления зверька и для нахождения зверька по id. Добвил новый класс для обработки исключения на случай, если зверек не был найден. Поменял свойство dll-auto на update.

+++ src/main/java/com/animals_back/service/AnimalService.java
@@ -2,6 +2,7 @@ package com.animals_back.service;
 
 import com.animals_back.dao.AnimalDAO;
 import com.animals_back.entity.Animal;
+import com.animals_back.exceptions.AnimalNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -21,11 +22,16 @@ public class AnimalService {
         animalDAO.save(animal);
     }
 
-    public Optional<Animal> findAnimalById(int id) {
-        return animalDAO.findById(id);
+    public Optional<Animal> findAnimalById(int id) throws AnimalNotFoundException {
+        return Optional.ofNullable(animalDAO.findById(id).orElseThrow(() -> new AnimalNotFoundException(id)));
     }
 
-    public void deleteAnimalById(int id) {
-        animalDAO.deleteById(id);
+    public void deleteAnimalById(int id) throws AnimalNotFoundException {
+        Optional<Animal> animal = this.findAnimalById(id);
+        if (animal.isPresent()) {
+            animalDAO.deleteById(id);
+        } else {
+            throw new AnimalNotFoundException(id);
+        }
     }
 }
